move  ViewName to be a global type
