+++ frontend/src/pages/operation_show/evidence_list/index.tsx
@@ -5,9 +5,8 @@ import * as React from 'react'
 import Layout from '../layout'
 import Timeline from 'src/components/timeline'
 import { EditEvidenceModal, DeleteEvidenceModal, ChangeFindingsOfEvidenceModal, MoveEvidenceModal } from '../evidence_modals'
-import { Evidence } from 'src/global_types'
+import { Evidence, ViewName } from 'src/global_types'
 import { RouteComponentProps } from 'react-router-dom'
-import { ViewName } from '../types'
 import { getEvidenceList } from 'src/services'
 import { useWiredData, useModal, renderModals } from 'src/helpers'
