--- frontend/src/pages/operation_show/types.ts
@@ -1,4 +1,0 @@
-// Copyright 2020, Verizon Media
-// Licensed under the terms of the MIT. See LICENSE file in project root for terms.
-
-export type ViewName = 'evidence' | 'findings'
